requirements.txt
fastapi>=0.135.0
uvicorn>=0.41.0
pydantic>=2.0
pydantic-settings>=2.0
sqlalchemy>=2.0
aiosqlite>=0.20.0
google-genai>=1.59.0
edge-tts>=7.2.7
qwen-tts>=0.1.1
requests>=2.31.0
numpy
soundfile
Pillow>=10.0
python-multipart
python-dotenv
